bit manipulation

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -12,6 +12,6 @@ int set_bit(unsigned long int *n, unsigned int index)
 	if (index > 63)
 		return (-1);
 
-	*n = ((unsigned long int) 1 << index) | *n;
+	*n = ((1ul << index) | *n);
 	return (1);
 }
